Move components to right location
